Created a mobile css sheet
and added basic styling to it. It looks good(ish) on mobile right now.

However, there is an issue with the nabber if the screen is not wide
enough. Most laptops would be ok, but horizontal tablets would suffer.
However, I cannot figure out a way that I could make it flow without
too much trouble.